Make gluestick generate coverage reports
1) Get coverage working with `test` command. For now, we are generating
html reports.

2) Extract karma config to separate file. I initially did this work as
part of creating a separate coverage command. I was going to revert it
but it seemed a little cleaner having the configuration separated from
the actual test command itself, so I left it as-is.
