+++ new/package.json
@@ -33,6 +33,7 @@
     "babel-plugin-react-transform": "1.1.1",
     "eslint": "^1.10.3",
     "eslint-plugin-react": "^3.15.0",
+    "isparta-loader": "1.0.0",
     "node-inspector": "0.12.6",
     "react-hot-loader": "1.3.0",
     "react-transform-catch-errors": "1.0.0",
