+++ package.json
@@ -49,6 +49,7 @@
     "karma": "0.13.19",
     "karma-chai": "0.1.0",
     "karma-chrome-launcher": "0.2.1",
+    "karma-coverage": "0.5.3",
     "karma-firefox-launcher": "0.1.7",
     "karma-mocha": "0.2.1",
     "karma-notify-reporter": "0.1.1",
