+++ src/commands/start-client.js
@@ -5,7 +5,7 @@ var chalk = require("chalk");
 var express = require("express");
 var proxy = require("express-http-proxy");
 var WebpackIsomorphicToolsPlugin = require("webpack-isomorphic-tools/plugin");
-var shared = require("./shared");
+var shared = require("../lib/shared");
 var detectEnvironmentVariables = require("../lib/detectEnvironmentVariables");
 var getWebpackAdditions = require("../lib/get-webpack-additions");
 var { additionalLoaders, additionalPreLoaders } = getWebpackAdditions();
