+++ src/lib/shared.js
